change from int to float

+++ js/script.js
@@ -10,9 +10,9 @@
  */
 function calculate () {
   // input
-  const length = parseInt(document.getElementById('length-of-pyramid').value)
-  const width = parseInt(document.getElementById('width-of-pyramid').value)
-  const height = parseInt(document.getElementById('height-of-pyramid').value)
+  const length = parseFloat(document.getElementById('length-of-pyramid').value)
+  const width = parseFloat(document.getElementById('width-of-pyramid').value)
+  const height = parseFloat(document.getElementById('height-of-pyramid').value)
 
   // process
   const volume = (length * width * height)  / 3
